feat: new api for resume data revert: all updated dependencies revert: ts-check for next.config.js
