+++ package.json
@@ -14,29 +14,30 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@prisma/client": "^4.1.1",
+    "@prisma/client": "^4.0.0",
     "@supabase/supabase-auth-helpers": "^1.4.2",
-    "@tailwindcss/typography": "^0.5.4",
+    "@tailwindcss/typography": "^0.5.2",
     "@types/uuid": "^8.3.4",
-    "autoprefixer": "^10.4.8",
+    "autoprefixer": "^10.4.7",
     "chart.js": "^3.8.0",
     "chrome-aws-lambda": "^10.1.0",
-    "cloc": "^2.10.0",
+    "cloc": "^2.9.0",
     "easymde": "^2.16.1",
     "feed": "^4.2.2",
     "formidable": "^2.0.1",
     "github-slugger": "^1.4.0",
     "html-escaper": "^3.0.3",
-    "next": "^12.2.3",
-    "next-superjson": "^0.0.4",
+    "image-size": "1.0.1",
+    "next": "^12.2.0",
+    "next-superjson": "^0.0.3",
     "next-themes": "^0.2.0",
     "postcss": "^8.4.14",
-    "preact": "^10.10.0",
+    "preact": "^10.8.2",
     "puppeteer-core": "^10.1.0",
     "react": "17.0.2",
-    "react-chartjs-2": "^4.3.1",
+    "react-chartjs-2": "^4.2.0",
     "react-dom": "17.0.2",
-    "react-hot-toast": "^2.3.0",
+    "react-hot-toast": "^2.2.0",
     "react-markdown": "^8.0.3",
     "react-select": "^5.4.0",
     "react-simplemde-editor": "^5.0.2",
@@ -44,7 +45,7 @@
     "sharp": "^0.30.7",
     "superjson": "^1.9.1",
     "swr": "1.3.0",
-    "tailwindcss": "^3.1.7",
+    "tailwindcss": "^3.1.4",
     "use-sound": "^4.0.1",
     "uuid": "^8.3.2",
     "web-vitals": "^2.1.4"
@@ -52,14 +53,14 @@
   "devDependencies": {
     "@commitlint/cli": "^17.0.3",
     "@commitlint/config-conventional": "^17.0.3",
-    "@next/bundle-analyzer": "12.2.3",
-    "@svgr/webpack": "^6.3.1",
+    "@next/bundle-analyzer": "12.2.0",
+    "@svgr/webpack": "^6.2.1",
     "@tailwindcss/forms": "^0.5.2",
     "@types/formidable": "^2.0.5",
     "@types/github-slugger": "^1.3.0",
     "@types/probe-image-size": "^7.2.0",
     "@types/react": "^18.0.14",
-    "@types/react-syntax-highlighter": "^15.5.4",
+    "@types/react-syntax-highlighter": "^15.5.2",
     "@types/ua-parser-js": "^0.7.36",
     "@typescript-eslint/eslint-plugin": "^5.30.3",
     "@typescript-eslint/parser": "^5.30.3",
@@ -75,11 +76,12 @@
     "gray-matter": "^4.0.3",
     "hastscript": "^7.0.2",
     "husky": "^8.0.1",
+    "inquirer": "^9.0.0",
     "lint-staged": "^13.0.3",
     "next-compose-plugins": "^2.2.1",
     "next-remote-watch": "^1.0.0",
     "prettier": "2.7.1",
-    "prisma": "^4.1.1",
+    "prisma": "^4.0.0",
     "typescript": "^4.7.4",
     "unist-util-visit": "^4.1.0"
   }
